+++ src/tools/consolidator/consolidator.ts
@@ -25,7 +25,11 @@ import {
 } from "./interactions/changeIssuer";
 import { validateMintNFT } from "./interactions/mintNFT";
 import { InMemoryAdapter } from "./adapters/in-memory-adapter";
-import {IConsolidatorAdapter} from "./adapters/types";
+import { IConsolidatorAdapter } from "./adapters/types";
+import {
+  consolidatedCollectionToInstance,
+  consolidatedNFTtoInstance,
+} from "./utils";
 
 export type ConsolidatorReturnType = {
   nfts: NFTConsolidated[];
@@ -64,72 +68,6 @@ export interface CollectionConsolidated {
   loadedMetadata?: CollectionMetadata;
   updatedAtBlock?: number;
 }
-
-export const consolidatedNFTtoInstance = (
-  nft?: NFTConsolidated
-): NFT | undefined => {
-  if (!nft) {
-    return undefined;
-  }
-  const {
-    block,
-    collection,
-    name,
-    instance,
-    transferable,
-    sn,
-    metadata,
-    id,
-    data,
-    updatedAtBlock,
-    ...rest
-  } = nft || {};
-  const nftClass = new NFT(
-    block,
-    collection,
-    name,
-    instance,
-    transferable,
-    sn,
-    metadata,
-    data,
-    updatedAtBlock
-  );
-  const { owner, forsale, reactions, changes, loadedMetadata, burned } = rest;
-  nftClass.owner = owner;
-  nftClass.forsale = forsale;
-  nftClass.reactions = reactions;
-  nftClass.changes = changes;
-  nftClass.loadedMetadata = loadedMetadata;
-  nftClass.burned = burned;
-
-  return nftClass;
-};
-
-export const consolidatedCollectionToInstance = (
-  collection?: CollectionConsolidated
-): Collection | undefined => {
-  if (!collection) {
-    return undefined;
-  }
-  const { block, name, metadata, id, issuer, max, symbol, ...rest } =
-    collection || {};
-  const colleactionClass = new Collection(
-    block,
-    name,
-    max,
-    issuer,
-    symbol,
-    id,
-    metadata
-  );
-  const { changes, loadedMetadata } = rest;
-
-  colleactionClass.changes = changes;
-  colleactionClass.loadedMetadata = loadedMetadata;
-
-  return colleactionClass;
-};
 
 export class Consolidator {
   readonly invalidCalls: InvalidCall[];
@@ -472,9 +410,7 @@ export class Consolidator {
       changeIssuerEntity.id
     );
 
-    const collection = consolidatedCollectionToInstance(
-      consolidatedCollection
-    );
+    const collection = consolidatedCollectionToInstance(consolidatedCollection);
 
     try {
       changeIssuerInteraction(remark, changeIssuerEntity, collection);
