+++ src/tools/consolidator/utils.ts
@@ -1,0 +1,69 @@
+import { NFT } from "../../rmrk1.0.0/classes/nft";
+import { CollectionConsolidated, NFTConsolidated } from "./consolidator";
+import { Collection } from "../../rmrk1.0.0/classes/collection";
+
+export const consolidatedNFTtoInstance = (
+  nft?: NFTConsolidated
+): NFT | undefined => {
+  if (!nft) {
+    return undefined;
+  }
+  const {
+    block,
+    collection,
+    name,
+    instance,
+    transferable,
+    sn,
+    metadata,
+    id,
+    data,
+    updatedAtBlock,
+    ...rest
+  } = nft || {};
+  const nftClass = new NFT(
+    block,
+    collection,
+    name,
+    instance,
+    transferable,
+    sn,
+    metadata,
+    data,
+    updatedAtBlock
+  );
+  const { owner, forsale, reactions, changes, loadedMetadata, burned } = rest;
+  nftClass.owner = owner;
+  nftClass.forsale = forsale;
+  nftClass.reactions = reactions;
+  nftClass.changes = changes;
+  nftClass.loadedMetadata = loadedMetadata;
+  nftClass.burned = burned;
+
+  return nftClass;
+};
+
+export const consolidatedCollectionToInstance = (
+  collection?: CollectionConsolidated
+): Collection | undefined => {
+  if (!collection) {
+    return undefined;
+  }
+  const { block, name, metadata, id, issuer, max, symbol, ...rest } =
+    collection || {};
+  const colleactionClass = new Collection(
+    block,
+    name,
+    max,
+    issuer,
+    symbol,
+    id,
+    metadata
+  );
+  const { changes, loadedMetadata } = rest;
+
+  colleactionClass.changes = changes;
+  colleactionClass.loadedMetadata = loadedMetadata;
+
+  return colleactionClass;
+};
